Added interruption handler to main (#175)
* Added interruption handler to main

* Added heading for sugar launch execution
