+++ src/deploy/process.rs
@@ -9,7 +9,15 @@ use console::style;
 use futures::future::select_all;
 use rand::rngs::OsRng;
 use spl_associated_token_account::get_associated_token_address;
-use std::{cmp, collections::HashSet, str::FromStr, sync::Arc};
+use std::{
+    cmp,
+    collections::HashSet,
+    str::FromStr,
+    sync::{
+        atomic::{AtomicBool, Ordering},
+        Arc,
+    },
+};
 
 use mpl_candy_machine::accounts as nft_accounts;
 use mpl_candy_machine::instruction as nft_instruction;
@@ -209,6 +217,7 @@ pub async fn process_deploy(args: DeployArgs) -> Result<()> {
                 candy_pubkey,
                 &mut cache,
                 config_lines,
+                args.handler,
             )
             .await?;
 
@@ -426,6 +435,7 @@ async fn upload_config_lines(
     candy_pubkey: Pubkey,
     cache: &mut Cache,
     config_lines: Vec<Vec<(u32, ConfigLine)>>,
+    handler: Arc<AtomicBool>,
 ) -> Result<Vec<DeployError>> {
     println!(
         "Sending config line(s) in {} transaction(s): (Ctrl+C to abort)",
@@ -461,7 +471,7 @@ async fn upload_config_lines(
 
     let mut errors = Vec::new();
 
-    while !handles.is_empty() {
+    while handler.load(Ordering::SeqCst) && !handles.is_empty() {
         match select_all(handles).await {
             (Ok(res), _index, remaining) => {
                 // independently if the upload was successful or not
